refactor: introduce method class for NFA converting
